Make System.Data.DataSetExtensions inbox, remove package
We pushed all the types down into System.Data.Common, and the facade is
now inbox, so the package only remains meaningful to carry a netstandard2.0
implementation assembly.  This assembly is only applicable on older
netcoreapp versions, so rather than ship it from master we'll  service
it from the release/2.1 branch.
