+++ pkg/Microsoft.Private.PackageBaseline/packageIndex.json
@@ -1232,7 +1232,9 @@
       ],
       "BaselineVersion": "4.5.0",
       "InboxOn": {
-        "net45": "4.0.0.0"
+        "net45": "4.0.0.0",
+        "netcoreapp3.0": "4.0.0.0",
+        "uap10.0.16300": "4.0.0.0"
       },
       "AssemblyVersionInPackageVersion": {
         "4.0.0.0": "4.6.0"
