+++ src/System.Data.Common/src/System.Data.Common.csproj
@@ -63,13 +63,9 @@
     <Compile Include="System\Data\DataRowChangeEvent.cs" />
     <Compile Include="System\Data\DataRowChangeEventHandler.cs" />
     <Compile Include="System\Data\DataRowCollection.cs" />
-    <Compile Include="$(CommonPath)\System\Data\DataRowComparer.cs">
-      <Link>System\Data\DataRowComparer.cs</Link>
-    </Compile>
+    <Compile Include="System\Data\DataRowComparer.cs" />
     <Compile Include="System\Data\DataRowCreatedEventHandler.cs" />
-    <Compile Include="$(CommonPath)\System\Data\DataRowExtensions.cs">
-      <Link>System\Data\DataRowExtensions.cs</Link>
-    </Compile>
+    <Compile Include="System\Data\DataRowExtensions.cs" />
     <Compile Include="System\Data\DataRowState.cs" />
     <Compile Include="System\Data\DataRowVersion.cs" />
     <Compile Include="System\Data\DataRowView.cs" />
@@ -77,18 +73,13 @@
     <Compile Include="System\Data\DataSet.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="$(CommonPath)\System\Data\DataSetUtil.cs">
-      <Link>System\Data\DataSetUtil.cs</Link>
-    </Compile>
+    <Compile Include="System\Data\DataSetUtil.cs" />
     <Compile Include="System\Data\DataSetDateTime.cs" />
     <Compile Include="System\Data\DataSysAttribute.cs" />
     <Compile Include="System\Data\DataTable.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="$(CommonPath)\System\Data\DataTableExtensions.cs">
-      <Link>System\Data\DataTableExtensions.cs</Link>
-    </Compile>
-    <Compile Include="System\Data\DataTableExtensions.AsDataView.cs" />
+    <Compile Include="System\Data\DataTableExtensions.cs" />
     <Compile Include="System\Data\DataTableClearEvent.cs" />
     <Compile Include="System\Data\DataTableClearEventHandler.cs" />
     <Compile Include="System\Data\DataTableCollection.cs" />
@@ -112,13 +103,8 @@
     <Compile Include="System\Data\DBConcurrencyException.cs" />
     <Compile Include="System\Data\DbType.cs" />
     <Compile Include="System\Data\DefaultValueTypeConverter.cs" />
-    <Compile Include="$(CommonPath)\System\Data\EnumerableRowCollection.cs">
-      <Link>System\Data\EnumerableRowCollection.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\EnumerableRowCollectionExtensions.cs">
-      <Link>System\Data\EnumerableRowCollectionExtensions.cs</Link>
-    </Compile>
-    <Compile Include="System\Data\EnumerableRowCollection.GetLinqDataView.cs" />
+    <Compile Include="System\Data\EnumerableRowCollection.cs" />
+    <Compile Include="System\Data\EnumerableRowCollectionExtensions.cs" />
     <Compile Include="System\Data\FillErrorEventArgs.cs" />
     <Compile Include="System\Data\FillErrorEventHandler.cs" />
     <Compile Include="System\Data\ForeignKeyConstraint.cs" />
@@ -146,9 +132,7 @@
     <Compile Include="System\Data\Merger.cs" />
     <Compile Include="System\Data\MissingMappingAction.cs" />
     <Compile Include="System\Data\MissingSchemaAction.cs" />
-    <Compile Include="$(CommonPath)\System\Data\OrderedEnumerableRowCollection.cs">
-      <Link>System\Data\OrderedEnumerableRowCollection.cs</Link>
-    </Compile>
+    <Compile Include="System\Data\OrderedEnumerableRowCollection.cs" />
     <Compile Include="System\Data\ParameterDirection.cs" />
     <Compile Include="System\Data\PrimaryKeyTypeConverter.cs" />
     <Compile Include="System\Data\PropertyCollection.cs" />
@@ -167,20 +151,14 @@
     <Compile Include="System\Data\Select.cs" />
     <Compile Include="System\Data\Selection.cs" />
     <Compile Include="System\Data\SimpleType.cs" />
-    <Compile Include="$(CommonPath)\System\Data\SortExpressionBuilder.cs">
-      <Link>System\Data\SortExpressionBuilder.cs</Link>
-    </Compile>
+    <Compile Include="System\Data\SortExpressionBuilder.cs" />
     <Compile Include="System\Data\SqlDbType.cs" />
     <Compile Include="System\Data\StateChangeEvent.cs" />
     <Compile Include="System\Data\StateChangeEventHandler.cs" />
     <Compile Include="System\Data\StatementType.cs" />
     <Compile Include="System\Data\StrongTypingException.cs" />
-    <Compile Include="$(CommonPath)\System\Data\TypedTableBase.cs">
-      <Link>System\Data\TypedTableBaseExtensions.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\TypedTableBaseExtensions.cs">
-      <Link>System\Data\TypedTableBaseExtensions.cs</Link>
-    </Compile>
+    <Compile Include="System\Data\TypedTableBase.cs" />
+    <Compile Include="System\Data\TypedTableBaseExtensions.cs" />
     <Compile Include="System\Data\UniqueConstraint.cs" />
     <Compile Include="System\Data\UpdateRowSource.cs" />
     <Compile Include="System\Data\Common\UInt64Storage.cs" />
