+++ src/System.Data.Common/src/System/Data/DataRowComparer.cs
