+++ src/System.Data.Common/src/System/Data/EnumerableRowCollectionExtensions.cs
