+++ src/System.Data.Common/src/System/Data/TypedTableBase.cs
