+++ src/System.Data.DataSetExtensions/Directory.Build.props
@@ -1,7 +1,10 @@
 ﻿<Project>
   <Import Project="..\Directory.Build.props" />
   <PropertyGroup>
+    <!-- AssemblyVersion is frozen to that which originally shipped in 2.1 -->
     <AssemblyVersion>4.0.0.0</AssemblyVersion>
     <StrongNameKeyId>ECMA</StrongNameKeyId>
+    <IsNETCoreApp>true</IsNETCoreApp>
+    <IsUAP>true</IsUAP>
   </PropertyGroup>
 </Project>
