--- src/System.Data.DataSetExtensions/pkg/System.Data.DataSetExtensions.pkgproj
@@ -1,13 +1,0 @@
-﻿<Project DefaultTargets="Build">
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), Directory.Build.props))\Directory.Build.props" />
-  <ItemGroup>
-    <ProjectReference Include="..\ref\System.Data.DataSetExtensions.csproj">
-      <SupportedFramework>uap10.0.16299;netcoreapp2.0;net45;$(AllXamarinFrameworks)</SupportedFramework>
-    </ProjectReference>
-    <ProjectReference Include="..\src\System.Data.DataSetExtensions.csproj" />
-    <InboxOnTargetFramework Include="net45">
-      <AsFrameworkReference>true</AsFrameworkReference>
-    </InboxOnTargetFramework>
-  </ItemGroup>
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), Directory.Build.targets))\Directory.Build.targets" />
-</Project>
