+++ src/System.Data.DataSetExtensions/src/Configurations.props
@@ -1,10 +1,8 @@
 ﻿<Project>
   <PropertyGroup>
     <BuildConfigurations>
-      netstandard;
       netcoreapp;
       uap;
-      _netfx;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
