--- src/System.Data.DataSetExtensions/src/Resources/Strings.resx
@@ -1,81 +1,0 @@
-﻿<root>
-  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
-    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
-    <xsd:element name="root" msdata:IsDataSet="true">
-      <xsd:complexType>
-        <xsd:choice maxOccurs="unbounded">
-          <xsd:element name="metadata">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" />
-              </xsd:sequence>
-              <xsd:attribute name="name" use="required" type="xsd:string" />
-              <xsd:attribute name="type" type="xsd:string" />
-              <xsd:attribute name="mimetype" type="xsd:string" />
-              <xsd:attribute ref="xml:space" />
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="assembly">
-            <xsd:complexType>
-              <xsd:attribute name="alias" type="xsd:string" />
-              <xsd:attribute name="name" type="xsd:string" />
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="data">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
-                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
-              </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
-              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
-              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
-              <xsd:attribute ref="xml:space" />
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="resheader">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
-              </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" />
-            </xsd:complexType>
-          </xsd:element>
-        </xsd:choice>
-      </xsd:complexType>
-    </xsd:element>
-  </xsd:schema>
-  <resheader name="resmimetype">
-    <value>text/microsoft-resx</value>
-  </resheader>
-  <resheader name="version">
-    <value>2.0</value>
-  </resheader>
-  <resheader name="reader">
-    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <resheader name="writer">
-    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </resheader>
-  <data name="DataSetLinq_InvalidEnumerationValue" xml:space="preserve">
-    <value>The {0} enumeration value, {1}, is not valid.</value>
-  </data>
-  <data name="DataSetLinq_EmptyDataRowSource" xml:space="preserve">
-    <value>The source contains no DataRows.</value>
-  </data>
-  <data name="DataSetLinq_NullDataRow" xml:space="preserve">
-    <value>The source contains a DataRow reference that is null.</value>
-  </data>
-  <data name="DataSetLinq_CannotLoadDetachedRow" xml:space="preserve">
-    <value>The source contains a detached DataRow that cannot be copied to the DataTable.</value>
-  </data>
-  <data name="DataSetLinq_CannotCompareDeletedRow" xml:space="preserve">
-    <value>The DataRowComparer does not work with DataRows that have been deleted since it only compares current values.</value>
-  </data>
-  <data name="DataSetLinq_CannotLoadDeletedRow" xml:space="preserve">
-    <value>The source contains a deleted DataRow that cannot be copied to the DataTable.</value>
-  </data>
-  <data name="DataSetLinq_NonNullableCast" xml:space="preserve">
-    <value>Cannot cast DBNull. Value to type '{0}'. Please use a nullable type.</value>
-  </data>
-</root>
