+++ src/System.Data.DataSetExtensions/src/System.Data.DataSetExtensions.csproj
@@ -1,11 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <ProjectGuid>{C7CB4B69-2A11-4A20-A21E-5C954855AE5A}</ProjectGuid>
     <AssemblyName>System.Data.DataSetExtensions</AssemblyName>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <Configurations>netstandard-Debug;netstandard-Release</Configurations>
-    <IsPartialFacadeAssembly Condition="'$(TargetsNETStandard)' != 'true'">true</IsPartialFacadeAssembly>
-    <OmitResources Condition="'$(IsPartialFacadeAssembly)' == 'true'">true</OmitResources>
+    <Configurations>netcoreapp-Debug;netcoreapp-Release;uap-Debug;uap-Release</Configurations>
+    <IsPartialFacadeAssembly>true</IsPartialFacadeAssembly>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System.Collections" />
@@ -16,37 +13,5 @@
     <Reference Include="System.Data.Common" />
     <Reference Include="System.Linq" />
     <Reference Include="System.ComponentModel.TypeConverter" />
-  </ItemGroup>
-  <ItemGroup Condition="'$(TargetsNETStandard)' == 'true'">
-    <Compile Include="$(CommonPath)\System\Data\DataRowComparer.cs">
-      <Link>System\Data\DataRowComparer.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\DataRowExtensions.cs">
-      <Link>System\Data\DataRowExtensions.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\DataSetUtil.cs">
-      <Link>System\Data\DataSetUtil.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\DataTableExtensions.cs">
-      <Link>System\Data\DataTableExtensions.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\EnumerableRowCollection.cs">
-      <Link>System\Data\EnumerableRowCollection.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\EnumerableRowCollectionExtensions.cs">
-      <Link>System\Data\EnumerableRowCollectionExtensions.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\OrderedEnumerableRowCollection.cs">
-      <Link>System\Data\OrderedEnumerableRowCollection.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\SortExpressionBuilder.cs">
-      <Link>System\Data\SortExpressionBuilder.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\TypedTableBase.cs">
-      <Link>System\Data\SortExpressionBuilder.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\System\Data\TypedTableBaseExtensions.cs">
-      <Link>System\Data\SortExpressionBuilder.cs</Link>
-    </Compile>
   </ItemGroup>
 </Project>
